Moves IsReferencing() to AssemblyExtensions
